refactor(studio): share one cell indexer across the gain-table volumes
The three static volume providers — per-speaker heatmap, global energy and
discontinuity — each carried their own copy of `sigEqual`, `clampIdx`,
`nearestIndex` and the `lookupZi`/`cellIndex` pair. Four identical helpers,
three times, with the non-obvious part (the height axis is non-uniform, so its
index comes from the table's real cell-centre positions, memoised because the
volume core walks depth innermost) explained in only one of them.

Move them to `object-energy-shared.js` as `signaturesEqual()` and
`makeCellIndexer(table)`, and document the height-axis reasoning there once.

Pure deduplication: same arithmetic, same memoisation, same signature
semantics, no behaviour change.

Context: this closes the Phase 1 slot of the Studio rebalancing plan. The two
CPU hotspots that phase was written against are not in the tree — the depth
warp is inverted once per depth slice, not per voxel (`energy-volume-core.js`
precomputes `depthByI`, n inversions rather than n³), and the height lookup is
already memoised across each row and gated behind a build signature. Neither
warranted the planned LUT.
