+++ omniphony-studio/src/scene/discontinuity-volume.js
@@ -22,7 +22,12 @@
 
 import { app } from '../state.js';
 import { EnergyVolume } from './energy-volume-core.js';
-import { VOLUME_REBUILD_INTERVAL_MS, clampVolumeGamma } from './object-energy-shared.js';
+import {
+  VOLUME_REBUILD_INTERVAL_MS,
+  clampVolumeGamma,
+  makeCellIndexer,
+  signaturesEqual,
+} from './object-energy-shared.js';
 import { getDiscontinuityTable } from './speaker-gaintable.js';
 
 const volume = new EnergyVolume();
@@ -33,13 +38,6 @@ export const DISCONTINUITY_SCALE_MAX = 2;
 export const DISCONTINUITY_SCALE_DEFAULT = 0.5;
 
 let lastBuildSig = null;
-function sigEqual(a, b) {
-  if (!a || !b || a.length !== b.length) return false;
-  for (let i = 0; i < a.length; i += 1) {
-    if (a[i] !== b[i]) return false;
-  }
-  return true;
-}
 
 export function hideDiscontinuityVolume() {
   volume.hide();
@@ -54,25 +52,6 @@ export function discontinuityScale() {
   const raw = Number(app.discontinuityHeatmapScale);
   if (!Number.isFinite(raw) || raw <= 0) return DISCONTINUITY_SCALE_DEFAULT;
   return Math.max(DISCONTINUITY_SCALE_MIN, Math.min(DISCONTINUITY_SCALE_MAX, raw));
-}
-
-function clampIdx(value, n) {
-  if (value < 0) return 0;
-  if (value > n - 1) return n - 1;
-  return value;
-}
-
-function nearestIndex(positions, n, value) {
-  let best = 0;
-  let bestDist = Infinity;
-  for (let i = 0; i < n; i += 1) {
-    const d = Math.abs(positions[i] - value);
-    if (d < bestDist) {
-      bestDist = d;
-      best = i;
-    }
-  }
-  return best;
 }
 
 export function refreshDiscontinuityVolume(nowMs) {
@@ -100,7 +79,7 @@ export function refreshDiscontinuityVolume(nowMs) {
     app.objectEnergyVolumeGammaMip,
     ratio.height, ratio.lower, ratio.width, ratio.rear, ratio.length,
   ];
-  if (sigEqual(sig, lastBuildSig)) return;
+  if (signaturesEqual(sig, lastBuildSig)) return;
 
   const now = Number.isFinite(nowMs) ? nowMs : performance.now();
   const refreshMs = Number(app.volumeRefreshMs) > 0
@@ -110,7 +89,7 @@ export function refreshDiscontinuityVolume(nowMs) {
   app.lastDiscontinuityVolumeAt = now;
   lastBuildSig = sig;
 
-  const { nx, ny, nz, bands, zPositions } = table;
+  const { bands } = table;
   const nbands = bands.length;
   if (nbands < 1) {
     hideDiscontinuityVolume();
@@ -134,25 +113,7 @@ export function refreshDiscontinuityVolume(nowMs) {
     }
     jumps = worst;
   }
-  const nxh = nx - 1;
-  const nyh = ny - 1;
-  const nzh = nz - 1;
-
-  let cachedOh = NaN;
-  let cachedZi = 0;
-  const lookupZi = (oh) => {
-    if (oh === cachedOh) return cachedZi;
-    cachedOh = oh;
-    cachedZi = zPositions
-      ? nearestIndex(zPositions, nz, oh)
-      : clampIdx(Math.round(((oh + 1) * 0.5) * nzh), nz);
-    return cachedZi;
-  };
-  const cellIndex = (ow, od, oh) => {
-    const xi = clampIdx(Math.round(((ow + 1) * 0.5) * nxh), nx);
-    const yi = clampIdx(Math.round(((od + 1) * 0.5) * nyh), ny);
-    return xi + nx * (yi + ny * lookupZi(oh));
-  };
+  const cellIndex = makeCellIndexer(table);
 
   const invScale = 1 / scale;
   volume.update({
